Move LocalSwapId to cnd
